requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
beanie==1.23.6
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
frozenlist==1.4.0
idna==3.6
lazy-model==0.2.0
loguru==0.7.2
magic-filter==1.0.12
motor==3.3.2
multidict==6.0.4
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dotenv==1.0.0
toml==0.10.2
typing_extensions==4.9.0
yarl==1.9.4
